refactor: major cleanup and file reorganization with sensible naming
Archive obsolete files:
- Move 5 older annotation scripts to archive/scripts/
- Move 3 database model versions to archive/database/
- Move 5 pipeline annotator implementations to archive/pipeline/
- Move 4 dashboard versions to archive/dashboard/

Rename for clarity:
- production_annotate_with_mft.py → annotate_interviews.py
- models_mft.py → models.py (consolidated)
- enhanced_json_annotator.py → json_annotator.py
- moral_foundations_analyzer.py → mft_analyzer.py
- holistic_chat_interface.py → chat_interface.py
- interactive_enhanced_app.py → app.py

Clean up:
- Remove empty placeholder docs (training/README, research/README)
- Move test_mft_annotation.py to tests directory
- Update all imports across 11 files to use new names

Result: Cleaner, more maintainable codebase with intuitive naming
