--- docs/research/README.md
@@ -1,20 +1,0 @@
-# Research Documentation
-
-This directory contains research documentation and findings from the Uruguay Active Listening project.
-
-## Structure
-
-- **Research Findings**: Analysis results and insights from interview data
-- **Methodology**: Research methods and analytical approaches
-- **Publications**: Academic papers and research outputs
-- **Data Analysis**: Statistical and qualitative analysis documentation
-
-## Planned Documentation
-
-- Research methodology documentation
-- Findings and insights reports
-- Academic publication drafts
-- Data analysis procedures
-- Quality assurance protocols
-
-*This directory is currently being populated as research progresses.*
