--- docs/training/README.md
@@ -1,20 +1,0 @@
-# Training Documentation
-
-This directory contains training materials and guides for the Uruguay Active Listening project.
-
-## Structure
-
-- **User Training**: Guides for system users and researchers
-- **Technical Training**: Developer and administrator documentation
-- **Quality Assurance**: Training materials for annotation quality
-- **Workflows**: Step-by-step process documentation
-
-## Planned Documentation
-
-- System user training guides
-- Annotation quality training materials
-- Technical setup and configuration guides
-- Research methodology training
-- Dashboard usage tutorials
-
-*This directory is currently being populated as training materials are developed.*
