+++ src/dashboard/app.py
@@ -1,6 +1,6 @@
 """
-Uruguay Interview Analysis Dashboard
-Interactive exploration of citizen consultation data
+Interactive Enhanced Uruguay Interview Analysis Dashboard
+Advanced research platform with dynamic filtering and linked visualizations
 """
 import streamlit as st
 import pandas as pd
@@ -9,533 +9,618 @@ import plotly.graph_objects as go
 from datetime import datetime
 from pathlib import Path
 import sys
+import numpy as np
 
 # Add project root to path
-sys.path.append(str(Path(__file__).parent.parent.parent))
+project_root = Path(__file__).parent.parent.parent
+sys.path.insert(0, str(project_root))
 
-from src.database.repository import InterviewRepository
-from src.database.connection import DatabaseConnection
-from src.database.models import Interview
-from src.config.config_loader import get_config
-from src.dashboard.conversation_view import show_conversation_flow
+from src.database.connection import get_db
+from src.database.models import (
+    Interview, ParticipantProfile, NarrativeFeatures, Priority, Turn,
+    TurnFunctionalAnalysis, TurnContentAnalysis, TurnEmotionalAnalysis
+)
 
 # Configure page
 st.set_page_config(
-    page_title="Uruguay Interview Analysis",
+    page_title="Uruguay Interview Analysis - Interactive Research Platform",
     page_icon="🇺🇾",
     layout="wide",
     initial_sidebar_state="expanded"
 )
 
-# Custom CSS for better styling
+# Enhanced CSS
 st.markdown("""
 <style>
-    .stTabs [data-baseweb="tab-list"] button [data-testid="stMarkdownContainer"] p {
-        font-size: 18px;
-    }
-    .priority-card {
-        background-color: #f0f2f6;
-        padding: 20px;
-        border-radius: 10px;
-        margin: 10px 0;
-    }
     .metric-card {
         background-color: #ffffff;
         padding: 15px;
         border-radius: 8px;
         box-shadow: 0 2px 4px rgba(0,0,0,0.1);
+        margin: 10px 0;
     }
+    .filter-section {
+        background-color: #f8f9fa;
+        padding: 15px;
+        border-radius: 8px;
+        margin: 10px 0;
+    }
+    .narrative-frame {
+        padding: 8px;
+        border-radius: 5px;
+        margin: 3px 0;
+        font-weight: bold;
+        text-align: center;
+    }
+    .frame-decline { background-color: #ffebee; color: #c62828; }
+    .frame-progress { background-color: #e8f5e8; color: #2e7d32; }
+    .frame-stagnation { background-color: #fff3e0; color: #ef6c00; }
+    .frame-mixed { background-color: #f3e5f5; color: #7b1fa2; }
+    .quote-card {
+        background-color: #f5f5f5;
+        padding: 12px;
+        border-left: 4px solid #2196f3;
+        margin: 8px 0;
+        border-radius: 0 8px 8px 0;
+    }
+    .confidence-high { border-left-color: #4caf50; }
+    .confidence-medium { border-left-color: #ff9800; }
+    .confidence-low { border-left-color: #f44336; }
+    .stSelectbox > div > div { background-color: white; }
 </style>
 """, unsafe_allow_html=True)
 
 
-@st.cache_resource
-def get_database_connection():
-    """Get database connection (cached)."""
-    config = get_config()
-    return DatabaseConnection(config.database.url)
-
-
-@st.cache_data(ttl=300)  # Cache for 5 minutes
-def load_interview_data():
-    """Load all interview data from database."""
-    db = get_database_connection()
-    repo = InterviewRepository(db)
+class InteractiveDashboard:
+    """Enhanced interactive dashboard with filtering and linked visualizations."""
     
-    with db.get_session() as session:
-        interviews = repo.get_all(session)
-        
-        # Extract interview basic info for selection (not ORM objects)
-        interview_list = []
-        for interview in interviews:
-            interview_list.append({
-                'id': interview.id,
-                'interview_id': interview.interview_id,
-                'date': interview.date,
-                'time': interview.time,
-                'location': interview.location,
-                'department': interview.department,
-                'participant_count': interview.participant_count,
-                'has_annotations': len(interview.annotations) > 0,
-                'has_turns': len(interview.turns) > 0 if hasattr(interview, 'turns') else False
-            })
-        
-        # Convert to dataframe for easier manipulation
-        data = []
-        if interviews:
-            for interview in interviews:
-                if interview.annotations:
-                    for annotation in interview.annotations:
-                        base_data = {
-                            'interview_id': interview.id,
-                            'date': interview.date,
-                            'location': interview.location,
-                            'department': interview.department,
-                            'participant_count': interview.participant_count,
-                            'overall_sentiment': annotation.overall_sentiment,
-                            'dominant_emotion': annotation.dominant_emotion,
-                            'confidence_score': annotation.confidence_score,
-                            'model_used': f"{annotation.model_provider}/{annotation.model_name}"
-                        }
-                        data.append(base_data)
-                else:
-                    # Add interview without annotation
-                    base_data = {
-                        'interview_id': interview.id,
-                        'date': interview.date,
-                        'location': interview.location,
-                        'department': interview.department,
-                        'participant_count': interview.participant_count,
-                        'overall_sentiment': None,
-                        'dominant_emotion': None,
-                        'confidence_score': 0.0,
-                        'model_used': None
-                    }
-                    data.append(base_data)
-        
-        # Load priorities
-        priorities = []
-        if interviews:
-            for interview in interviews:
-                for priority in interview.priorities:
-                    priorities.append({
-                        'interview_id': interview.id,
-                        'location': interview.location,
-                        'department': interview.department,
-                        'date': interview.date,
-                        'scope': priority.scope,
-                        'rank': priority.rank,
-                        'category': priority.category,
-                        'description': priority.description
-                    })
-        
-        # Load themes
-        themes = []
-        if interviews:
-            for interview in interviews:
-                for theme in interview.themes:
-                    themes.append({
-                        'interview_id': interview.id,
-                        'location': interview.location,
-                        'department': interview.department,
-                        'theme': theme.theme,
-                        'frequency': theme.frequency
-                    })
+    def __init__(self):
+        self.db = get_db()
+        self.setup_session_state()
     
-    return pd.DataFrame(data), pd.DataFrame(priorities), pd.DataFrame(themes), interview_list
-
-
-def show_overview_dashboard(df_interviews, df_priorities, df_themes):
-    """Display overview dashboard with key metrics."""
-    st.title("🇺🇾 Uruguay Citizen Consultation Analysis")
-    st.markdown("### Overview Dashboard")
-    
-    # Key metrics
-    col1, col2, col3, col4 = st.columns(4)
-    
-    with col1:
-        st.metric(
-            "Total Interviews",
-            len(df_interviews),
-            delta=None,
-            help="Number of citizen interviews conducted"
-        )
-    
-    with col2:
-        avg_confidence = df_interviews['confidence_score'].mean()
-        st.metric(
-            "Avg Confidence",
-            f"{avg_confidence:.2f}",
-            delta=None,
-            help="Average AI annotation confidence score"
-        )
-    
-    with col3:
-        total_priorities = len(df_priorities)
-        st.metric(
-            "Priorities Identified",
-            total_priorities,
-            delta=None,
-            help="Total number of priorities across all interviews"
-        )
-    
-    with col4:
-        unique_themes = df_themes['theme'].nunique() if not df_themes.empty else 0
-        st.metric(
-            "Unique Themes",
-            unique_themes,
-            delta=None,
-            help="Number of distinct themes identified"
-        )
-    
-    # Sentiment distribution
-    st.markdown("### Sentiment Analysis")
-    col1, col2 = st.columns(2)
-    
-    with col1:
-        if not df_interviews.empty:
-            sentiment_counts = df_interviews['overall_sentiment'].value_counts()
-            fig_sentiment = px.pie(
-                values=sentiment_counts.values,
-                names=sentiment_counts.index,
-                title="Overall Sentiment Distribution",
-                color_discrete_map={
-                    'positive': '#2ecc71',
-                    'negative': '#e74c3c',
-                    'neutral': '#95a5a6',
-                    'mixed': '#3498db'
-                }
-            )
-            st.plotly_chart(fig_sentiment, use_container_width=True)
-    
-    with col2:
-        if not df_interviews.empty and 'dominant_emotion' in df_interviews.columns:
-            emotion_counts = df_interviews['dominant_emotion'].value_counts().head(10)
-            fig_emotion = px.bar(
-                x=emotion_counts.values,
-                y=emotion_counts.index,
-                orientation='h',
-                title="Top Emotional Themes",
-                labels={'x': 'Count', 'y': 'Emotion'}
-            )
-            st.plotly_chart(fig_emotion, use_container_width=True)
-    
-    # Geographic distribution
-    st.markdown("### Geographic Distribution")
-    if not df_interviews.empty:
-        location_counts = df_interviews.groupby(['department', 'location']).size().reset_index(name='count')
-        fig_geo = px.treemap(
-            location_counts,
-            path=['department', 'location'],
-            values='count',
-            title="Interviews by Location"
-        )
-        st.plotly_chart(fig_geo, use_container_width=True)
-
-
-def show_priorities_analysis(df_priorities):
-    """Display priorities analysis dashboard."""
-    st.title("📊 Priorities Analysis")
-    
-    if df_priorities.empty:
-        st.warning("No priorities data available")
-        return
-    
-    # Filter by scope
-    scope_filter = st.selectbox(
-        "Select Priority Scope",
-        ["All", "National", "Local"],
-        key="priority_scope"
-    )
-    
-    if scope_filter != "All":
-        filtered_priorities = df_priorities[df_priorities['scope'] == scope_filter.lower()]
-    else:
-        filtered_priorities = df_priorities
-    
-    # Top categories
-    col1, col2 = st.columns(2)
-    
-    with col1:
-        category_counts = filtered_priorities['category'].value_counts().head(10)
-        fig_categories = px.bar(
-            x=category_counts.index,
-            y=category_counts.values,
-            title=f"Top {scope_filter} Priority Categories",
-            labels={'x': 'Category', 'y': 'Count'}
-        )
-        st.plotly_chart(fig_categories, use_container_width=True)
-    
-    with col2:
-        # Priority ranking distribution
-        rank_counts = filtered_priorities['rank'].value_counts().sort_index()
-        fig_ranks = px.line(
-            x=rank_counts.index,
-            y=rank_counts.values,
-            title="Priority Rank Distribution",
-            labels={'x': 'Rank', 'y': 'Count'},
-            markers=True
-        )
-        st.plotly_chart(fig_ranks, use_container_width=True)
-    
-    # Detailed priority view
-    st.markdown("### Priority Details")
-    
-    # Group by category and show top priorities
-    for category in category_counts.index[:5]:
-        with st.expander(f"{category.title()} Priorities"):
-            cat_priorities = filtered_priorities[filtered_priorities['category'] == category].sort_values('rank')
-            for _, priority in cat_priorities.head(5).iterrows():
-                st.markdown(f"**Rank {priority['rank']}** ({priority['scope']}): {priority['description']}")
-                st.markdown(f"*Location: {priority['location']}*")
-                st.divider()
-
-
-def show_themes_exploration(df_themes):
-    """Display themes exploration dashboard."""
-    st.title("🏷️ Themes Exploration")
-    
-    if df_themes.empty:
-        st.warning("No themes data available")
-        return
-    
-    # Theme frequency
-    theme_freq = df_themes.groupby('theme')['frequency'].sum().sort_values(ascending=False).head(20)
-    
-    fig_themes = px.bar(
-        x=theme_freq.values,
-        y=theme_freq.index,
-        orientation='h',
-        title="Most Frequent Themes",
-        labels={'x': 'Total Frequency', 'y': 'Theme'}
-    )
-    fig_themes.update_layout(height=600)
-    st.plotly_chart(fig_themes, use_container_width=True)
-    
-    # Theme co-occurrence network (simplified)
-    st.markdown("### Theme Distribution by Location")
-    
-    # Theme heatmap by department
-    theme_dept = df_themes.groupby(['department', 'theme'])['frequency'].sum().reset_index()
-    theme_pivot = theme_dept.pivot(index='theme', columns='department', values='frequency').fillna(0)
-    
-    # Select top themes for better visualization
-    top_themes = theme_freq.head(15).index
-    theme_pivot_top = theme_pivot.loc[theme_pivot.index.isin(top_themes)]
-    
-    fig_heatmap = px.imshow(
-        theme_pivot_top,
-        labels=dict(x="Department", y="Theme", color="Frequency"),
-        title="Theme Frequency Heatmap by Department",
-        aspect="auto"
-    )
-    fig_heatmap.update_layout(height=600)
-    st.plotly_chart(fig_heatmap, use_container_width=True)
-
-
-def show_interview_detail(interviews):
-    """Display detailed interview viewer."""
-    st.title("📄 Interview Detail Viewer")
-    
-    if not interviews:
-        st.warning("No interviews available")
-        return
-    
-    # Get database connection for this view
-    db = get_database_connection()
-    
-    # Interview selector  
-    interview_options = {f"{i['id']} - {i['location']} ({i['date']})": i['id'] for i in interviews}
-    selected_key = st.selectbox(
-        "Select an interview to view",
-        options=list(interview_options.keys())
-    )
-    
-    if selected_key:
-        interview_id = interview_options[selected_key]
-        
-        # Load the specific interview with all relationships in a new session
-        with db.get_session() as session:
-            from sqlalchemy.orm import joinedload
-            interview = session.query(Interview).options(
-                joinedload(Interview.annotations),
-                joinedload(Interview.priorities),
-                joinedload(Interview.themes),
-                joinedload(Interview.emotions)
-            ).filter(Interview.id == interview_id).first()
-            
-            if not interview:
-                st.error("Interview not found")
-                return
-            
-            # Extract all data while session is active
-            interview_data = {
-                'date': interview.date,
-                'time': interview.time,
-                'location': interview.location,
-                'department': interview.department,
-                'participant_count': interview.participant_count,
-                'word_count': interview.word_count,
-                'raw_text': interview.raw_text,
-                'annotations': [{
-                    'overall_sentiment': ann.overall_sentiment,
-                    'dominant_emotion': ann.dominant_emotion
-                } for ann in interview.annotations],
-                'priorities': [{
-                    'scope': p.scope,
-                    'rank': p.rank,
-                    'category': p.category,
-                    'description': p.description
-                } for p in interview.priorities],
-                'themes': [{
-                    'theme': t.theme,
-                    'frequency': t.frequency
-                } for t in interview.themes],
-                'emotions': [{
-                    'type': e.type,
-                    'intensity': e.intensity,
-                    'context': e.context
-                } for e in interview.emotions]
+    def setup_session_state(self):
+        """Initialize session state for filters and selections."""
+        if 'filters' not in st.session_state:
+            st.session_state.filters = {
+                'departments': [],
+                'age_ranges': [],
+                'narrative_frames': [],
+                'locality_sizes': [],
+                'confidence_threshold': 0.0,
+                'gender': [],
+                'selected_interviews': []
             }
         
-        # Now display the data outside the session
-        # Interview metadata
-        col1, col2, col3 = st.columns(3)
-        with col1:
-            st.markdown(f"**Date:** {interview_data['date']}")
-            st.markdown(f"**Time:** {interview_data['time']}")
-        with col2:
-            st.markdown(f"**Location:** {interview_data['location']}")
-            st.markdown(f"**Department:** {interview_data['department']}")
-        with col3:
-            st.markdown(f"**Participants:** {interview_data['participant_count']}")
-            if interview_data['word_count']:
-                st.markdown(f"**Word Count:** {interview_data['word_count']:,}")
-        
-        st.divider()
-        
-        # Annotation details
-        if interview_data['annotations']:
-            annotation = interview_data['annotations'][0]  # Use first annotation
+        if 'selected_interview' not in st.session_state:
+            st.session_state.selected_interview = None
+    
+    @st.cache_data(ttl=300)
+    def load_comprehensive_data(_self):
+        """Load all data with relationships."""
+        with _self.db.get_session() as session:
+            # Load interviews with relationships
+            interviews = session.query(Interview).all()
             
-            # Tabs for different aspects
-            tab1, tab2, tab3, tab4, tab5 = st.tabs(["Priorities", "Themes", "Emotions", "Conversation Flow", "Raw Text"])
-            
-            with tab1:
-                st.markdown("### National Priorities")
-                nat_priorities = [p for p in interview_data['priorities'] if p['scope'] == 'national']
-                for priority in sorted(nat_priorities, key=lambda x: x['rank']):
-                    st.markdown(f"**Rank {priority['rank']}:** {priority['category']}")
-                    st.markdown(f"{priority['description']}")
-                    st.divider()
+            # Build comprehensive dataset
+            data = []
+            for interview in interviews:
+                profile = interview.participant_profile
+                narrative = interview.narrative_features
+                priorities = interview.priorities
                 
-                st.markdown("### Local Priorities")
-                local_priorities = [p for p in interview_data['priorities'] if p['scope'] == 'local']
-                for priority in sorted(local_priorities, key=lambda x: x['rank']):
-                    st.markdown(f"**Rank {priority['rank']}:** {priority['category']}")
-                    st.markdown(f"{priority['description']}")
-                    st.divider()
-            
-            with tab2:
-                st.markdown("### Identified Themes")
-                for theme in interview_data['themes']:
-                    st.markdown(f"- **{theme['theme']}** (frequency: {theme['frequency']})")
-            
-            with tab3:
-                st.markdown("### Emotional Analysis")
-                st.markdown(f"**Overall Sentiment:** {annotation['overall_sentiment']}")
-                st.markdown(f"**Dominant Emotion:** {annotation['dominant_emotion']}")
+                # Count priorities by scope
+                national_priorities = len([p for p in priorities if p.scope == 'national'])
+                local_priorities = len([p for p in priorities if p.scope == 'local'])
                 
-                if interview_data['emotions']:
-                    st.markdown("### Emotional Expressions")
-                    for emotion in interview_data['emotions']:
-                        st.markdown(f"- **{emotion['type']}** (intensity: {emotion['intensity']})")
-                        if emotion.get('context'):
-                            st.markdown(f"  Context: {emotion['context']}")
+                # Average emotional intensity
+                avg_emotional_intensity = np.mean([p.emotional_intensity for p in priorities if p.emotional_intensity]) if priorities else None
+                
+                # Top priority themes
+                priority_themes = [p.theme for p in priorities] if priorities else []
+                
+                data.append({
+                    'interview_id': interview.interview_id,
+                    'date': interview.date,
+                    'location': interview.location,
+                    'department': interview.department,
+                    'municipality': interview.municipality,
+                    'locality_size': interview.locality_size,
+                    'duration_minutes': interview.duration_minutes,
+                    'interviewer_ids': interview.interviewer_ids,
+                    
+                    # Participant profile
+                    'age_range': profile.age_range if profile else None,
+                    'gender': profile.gender if profile else None,
+                    'occupation_sector': profile.occupation_sector if profile else None,
+                    'organizational_affiliation': profile.organizational_affiliation if profile else None,
+                    'profile_confidence': profile.profile_confidence if profile else None,
+                    
+                    # Narrative features
+                    'dominant_frame': narrative.dominant_frame if narrative else None,
+                    'temporal_orientation': narrative.temporal_orientation if narrative else None,
+                    'government_responsibility': narrative.government_responsibility if narrative else None,
+                    'individual_responsibility': narrative.individual_responsibility if narrative else None,
+                    'structural_factors': narrative.structural_factors if narrative else None,
+                    'solution_orientation': narrative.solution_orientation if narrative else None,
+                    'cultural_patterns': narrative.cultural_patterns_identified if narrative else [],
+                    'narrative_confidence': narrative.narrative_confidence if narrative else None,
+                    
+                    # Priority aggregates
+                    'national_priorities': national_priorities,
+                    'local_priorities': local_priorities,
+                    'total_priorities': len(priorities),
+                    'avg_emotional_intensity': avg_emotional_intensity,
+                    'priority_themes': priority_themes,
+                })
             
-            with tab4:
-                # Show conversation flow
-                try:
-                    db = get_database_connection()
-                    with db.get_session() as conv_session:
-                        # Check if interview has turns
-                        from src.database.models import Turn
-                        
-                        turn_count = conv_session.query(Turn).filter_by(interview_id=interview_id).count()
-                        
-                        if turn_count > 0:
-                            show_conversation_flow(interview_id, conv_session)
-                        else:
-                            st.info("No conversation turns available for this interview")
-                except ImportError as e:
-                    st.error(f"Import error: {e}")
-                    st.info("Conversation flow feature requires the latest database models")
-                except Exception as e:
-                    st.error(f"Error loading conversation flow: {str(e)}")
-                    st.info("Conversation flow feature requires turn-level data")
+            # Load priority details
+            priorities_data = []
+            for interview in interviews:
+                for priority in interview.priorities:
+                    priorities_data.append({
+                        'interview_id': interview.interview_id,
+                        'scope': priority.scope,
+                        'rank': priority.rank,
+                        'theme': priority.theme,
+                        'specific_issues': priority.specific_issues,
+                        'narrative_elaboration': priority.narrative_elaboration,
+                        'emotional_intensity': priority.emotional_intensity,
+                        'supporting_quotes': priority.supporting_quotes,
+                        'confidence': priority.confidence,
+                        'reasoning': priority.reasoning
+                    })
             
-            with tab5:
-                st.markdown("### Interview Text")
-                if interview_data['raw_text']:
-                    st.text_area(
-                        "Full Interview Transcript",
-                        value=interview_data['raw_text'],
-                        height=400,
-                        disabled=True
-                    )
+            return pd.DataFrame(data), pd.DataFrame(priorities_data)
+    
+    def create_filter_sidebar(self, df):
+        """Create interactive filter sidebar."""
+        st.sidebar.title("🔍 Interactive Filters")
+        st.sidebar.markdown("Filter the data to explore patterns")
+        
+        # Reset filters button
+        if st.sidebar.button("🔄 Reset All Filters"):
+            for key in st.session_state.filters:
+                if isinstance(st.session_state.filters[key], list):
+                    st.session_state.filters[key] = []
                 else:
-                    st.info("Raw text not available in database")
+                    st.session_state.filters[key] = 0.0
+        
+        # Department filter
+        departments = df['department'].dropna().unique()
+        st.session_state.filters['departments'] = st.sidebar.multiselect(
+            "🗺️ Departments",
+            options=sorted(departments),
+            default=st.session_state.filters['departments'],
+            help="Select one or more departments"
+        )
+        
+        # Age range filter
+        age_ranges = df['age_range'].dropna().unique()
+        st.session_state.filters['age_ranges'] = st.sidebar.multiselect(
+            "👥 Age Ranges", 
+            options=sorted(age_ranges),
+            default=st.session_state.filters['age_ranges']
+        )
+        
+        # Gender filter
+        genders = df['gender'].dropna().unique()
+        st.session_state.filters['gender'] = st.sidebar.multiselect(
+            "⚧ Gender",
+            options=sorted(genders),
+            default=st.session_state.filters['gender']
+        )
+        
+        # Narrative frame filter
+        frames = df['dominant_frame'].dropna().unique()
+        st.session_state.filters['narrative_frames'] = st.sidebar.multiselect(
+            "📚 Narrative Frames",
+            options=sorted(frames),
+            default=st.session_state.filters['narrative_frames']
+        )
+        
+        # Locality size filter
+        localities = df['locality_size'].dropna().unique()
+        st.session_state.filters['locality_sizes'] = st.sidebar.multiselect(
+            "🏘️ Locality Size",
+            options=sorted(localities),
+            default=st.session_state.filters['locality_sizes']
+        )
+        
+        # Confidence threshold
+        st.session_state.filters['confidence_threshold'] = st.sidebar.slider(
+            "🎯 Minimum Confidence",
+            min_value=0.0,
+            max_value=1.0,
+            value=st.session_state.filters['confidence_threshold'],
+            step=0.1,
+            help="Filter by annotation confidence level"
+        )
+        
+        # Show filter summary
+        active_filters = sum(1 for f in st.session_state.filters.values() if f)
+        if active_filters > 0:
+            st.sidebar.success(f"✅ {active_filters} filters active")
+        else:
+            st.sidebar.info("ℹ️ No filters applied")
+    
+    def apply_filters(self, df):
+        """Apply active filters to dataframe."""
+        filtered_df = df.copy()
+        
+        filters = st.session_state.filters
+        
+        # Apply department filter
+        if filters['departments']:
+            filtered_df = filtered_df[filtered_df['department'].isin(filters['departments'])]
+        
+        # Apply age range filter
+        if filters['age_ranges']:
+            filtered_df = filtered_df[filtered_df['age_range'].isin(filters['age_ranges'])]
+        
+        # Apply gender filter
+        if filters['gender']:
+            filtered_df = filtered_df[filtered_df['gender'].isin(filters['gender'])]
+        
+        # Apply narrative frame filter
+        if filters['narrative_frames']:
+            filtered_df = filtered_df[filtered_df['dominant_frame'].isin(filters['narrative_frames'])]
+        
+        # Apply locality size filter
+        if filters['locality_sizes']:
+            filtered_df = filtered_df[filtered_df['locality_size'].isin(filters['locality_sizes'])]
+        
+        # Apply confidence threshold
+        if filters['confidence_threshold'] > 0:
+            filtered_df = filtered_df[
+                (filtered_df['narrative_confidence'] >= filters['confidence_threshold']) |
+                (filtered_df['narrative_confidence'].isna())
+            ]
+        
+        return filtered_df
+    
+    def show_overview_metrics(self, df, filtered_df):
+        """Enhanced overview metrics with filtering impact."""
+        col1, col2, col3, col4 = st.columns(4)
+        
+        with col1:
+            total = len(df)
+            filtered = len(filtered_df)
+            st.metric(
+                "Interviews", 
+                filtered,
+                delta=f"{filtered-total}" if filtered != total else None,
+                help=f"Filtered from {total} total interviews"
+            )
+        
+        with col2:
+            profiles = filtered_df['profile_confidence'].notna().sum()
+            st.metric("Participant Profiles", profiles)
+        
+        with col3:
+            narratives = filtered_df['narrative_confidence'].notna().sum()
+            st.metric("Narrative Features", narratives)
+        
+        with col4:
+            avg_conf = filtered_df['narrative_confidence'].mean()
+            st.metric(
+                "Avg Confidence", 
+                f"{avg_conf:.2f}" if pd.notna(avg_conf) else "N/A"
+            )
+        
+        # Show filter impact
+        if len(filtered_df) != len(df):
+            filter_pct = (len(filtered_df) / len(df)) * 100
+            st.info(f"📊 Showing {filter_pct:.1f}% of data ({len(filtered_df)}/{len(df)} interviews) based on active filters")
+    
+    def create_linked_narrative_analysis(self, filtered_df):
+        """Interactive narrative analysis with click-to-filter."""
+        st.subheader("📚 Interactive Narrative Frame Analysis")
+        
+        frames = filtered_df['dominant_frame'].value_counts()
+        
+        if not frames.empty:
+            col1, col2 = st.columns([2, 1])
+            
+            with col1:
+                # Interactive pie chart
+                fig = px.pie(
+                    values=frames.values,
+                    names=frames.index,
+                    title="Narrative Frames Distribution (Click to Filter)",
+                    color_discrete_map={
+                        'decline': '#ffcdd2',
+                        'progress': '#c8e6c9',
+                        'stagnation': '#ffe0b2', 
+                        'mixed': '#e1bee7',
+                        'progress and decline': '#d1c4e9'
+                    }
+                )
+                
+                # Add click event handling hint
+                fig.update_layout(
+                    title_text="Narrative Frames Distribution<br><sub>Click slice to filter dashboard</sub>"
+                )
+                
+                selected_points = st.plotly_chart(fig, use_container_width=True, key="narrative_frames_pie")
+            
+            with col2:
+                st.markdown("**Frame Details:**")
+                for frame, count in frames.items():
+                    frame_class = f"frame-{frame.replace(' ', '-').replace('and', '').strip().lower()}"
+                    pct = (count / len(filtered_df)) * 100
+                    
+                    # Make frames clickable to update filters
+                    if st.button(f"{frame.title()}: {count} ({pct:.1f}%)", key=f"frame_{frame}"):
+                        st.session_state.filters['narrative_frames'] = [frame]
+                        st.rerun()
+                    
+                    st.markdown(f'<div class="narrative-frame {frame_class}">{frame.title()}: {count}</div>', 
+                               unsafe_allow_html=True)
+    
+    def create_interactive_agency_analysis(self, filtered_df):
+        """Interactive agency attribution with filtering."""
+        st.subheader("🎭 Agency Attribution Analysis")
+        
+        # Filter out null values
+        agency_df = filtered_df[['government_responsibility', 'individual_responsibility', 'structural_factors']].dropna()
+        
+        if not agency_df.empty:
+            col1, col2 = st.columns([2, 1])
+            
+            with col1:
+                # Interactive box plot
+                fig = go.Figure()
+                
+                fig.add_trace(go.Box(
+                    y=agency_df['government_responsibility'],
+                    name='Government',
+                    marker_color='lightblue',
+                    boxpoints='outliers'
+                ))
+                
+                fig.add_trace(go.Box(
+                    y=agency_df['individual_responsibility'],
+                    name='Individual', 
+                    marker_color='lightgreen',
+                    boxpoints='outliers'
+                ))
+                
+                fig.add_trace(go.Box(
+                    y=agency_df['structural_factors'],
+                    name='Structural',
+                    marker_color='lightcoral',
+                    boxpoints='outliers'
+                ))
+                
+                fig.update_layout(
+                    title="Agency Attribution Distribution",
+                    yaxis_title="Responsibility Score (0-1)",
+                    showlegend=True
+                )
+                
+                st.plotly_chart(fig, use_container_width=True)
+            
+            with col2:
+                st.markdown("**Average Scores:**")
+                gov_avg = agency_df['government_responsibility'].mean()
+                ind_avg = agency_df['individual_responsibility'].mean()
+                str_avg = agency_df['structural_factors'].mean()
+                
+                st.metric("Government", f"{gov_avg:.2f}")
+                st.metric("Individual", f"{ind_avg:.2f}")
+                st.metric("Structural", f"{str_avg:.2f}")
+                
+                # Interpretation
+                st.markdown("**Interpretation:**")
+                if gov_avg > 0.7:
+                    st.success("🏛️ High government responsibility")
+                elif gov_avg > 0.5:
+                    st.warning("🏛️ Moderate government responsibility")
+                else:
+                    st.info("🏛️ Low government responsibility")
+    
+    def create_interview_selector(self, filtered_df):
+        """Interactive interview selection."""
+        st.subheader("📄 Interview Explorer")
+        
+        # Interview selection
+        interview_options = filtered_df['interview_id'].tolist()
+        
+        col1, col2 = st.columns([1, 2])
+        
+        with col1:
+            selected_interview = st.selectbox(
+                "Select Interview to Explore",
+                options=interview_options,
+                index=0 if interview_options else None,
+                help="Choose an interview for detailed analysis"
+            )
+        
+        with col2:
+            if selected_interview:
+                interview_data = filtered_df[filtered_df['interview_id'] == selected_interview].iloc[0]
+                
+                # Quick preview
+                col2a, col2b, col2c = st.columns(3)
+                with col2a:
+                    st.metric("Location", interview_data['location'])
+                with col2b:
+                    st.metric("Frame", interview_data['dominant_frame'])
+                with col2c:
+                    st.metric("Priorities", interview_data['total_priorities'])
+        
+        return selected_interview
+    
+    def show_detailed_interview_analysis(self, interview_id, df):
+        """Enhanced interview detail view."""
+        interview_data = df[df['interview_id'] == interview_id].iloc[0]
+        
+        st.markdown(f"### 🔍 Interview {interview_id} - Detailed Analysis")
+        
+        # Tabbed interface for different aspects
+        tab1, tab2, tab3, tab4 = st.tabs(["📋 Overview", "👤 Participant", "📚 Narrative", "🎯 Priorities"])
+        
+        with tab1:
+            col1, col2 = st.columns(2)
+            
+            with col1:
+                st.markdown("**Interview Metadata:**")
+                st.write(f"📅 Date: {interview_data['date']}")
+                st.write(f"📍 Location: {interview_data['location']}")
+                st.write(f"🗺️ Department: {interview_data['department']}")
+                st.write(f"🏘️ Municipality: {interview_data['municipality']}")
+                st.write(f"⏱️ Duration: {interview_data['duration_minutes']} minutes")
+                
+                if interview_data['interviewer_ids']:
+                    st.write(f"🎤 Interviewers: {', '.join(interview_data['interviewer_ids'])}")
+            
+            with col2:
+                st.markdown("**Analysis Summary:**")
+                st.write(f"🎯 Total Priorities: {interview_data['total_priorities']}")
+                st.write(f"🏛️ National Priorities: {interview_data['national_priorities']}")
+                st.write(f"🏠 Local Priorities: {interview_data['local_priorities']}")
+                
+                if interview_data['avg_emotional_intensity']:
+                    st.write(f"😊 Avg Emotional Intensity: {interview_data['avg_emotional_intensity']:.2f}")
+        
+        with tab2:
+            col1, col2 = st.columns(2)
+            
+            with col1:
+                st.markdown("**Demographics:**")
+                st.write(f"👥 Age Range: {interview_data['age_range']}")
+                st.write(f"⚧ Gender: {interview_data['gender']}")
+                st.write(f"💼 Occupation: {interview_data['occupation_sector']}")
+                
+                if interview_data['organizational_affiliation']:
+                    st.write(f"🏢 Affiliation: {interview_data['organizational_affiliation']}")
+            
+            with col2:
+                st.markdown("**Profile Quality:**")
+                if interview_data['profile_confidence']:
+                    confidence = interview_data['profile_confidence']
+                    st.metric("Profile Confidence", f"{confidence:.2f}")
+                    
+                    if confidence >= 0.8:
+                        st.success("🟢 High confidence profile")
+                    elif confidence >= 0.6:
+                        st.warning("🟡 Medium confidence profile") 
+                    else:
+                        st.error("🔴 Low confidence profile")
+        
+        with tab3:
+            col1, col2 = st.columns(2)
+            
+            with col1:
+                st.markdown("**Narrative Characteristics:**")
+                st.write(f"📖 Dominant Frame: {interview_data['dominant_frame']}")
+                st.write(f"⏰ Temporal Orientation: {interview_data['temporal_orientation']}")
+                st.write(f"🎯 Solution Orientation: {interview_data['solution_orientation']}")
+            
+            with col2:
+                st.markdown("**Agency Attribution:**")
+                if interview_data['government_responsibility']:
+                    st.metric("Government", f"{interview_data['government_responsibility']:.2f}")
+                if interview_data['individual_responsibility']:
+                    st.metric("Individual", f"{interview_data['individual_responsibility']:.2f}")
+                if interview_data['structural_factors']:
+                    st.metric("Structural", f"{interview_data['structural_factors']:.2f}")
+            
+            # Cultural patterns
+            if interview_data['cultural_patterns']:
+                st.markdown("**Cultural Patterns:**")
+                for pattern in interview_data['cultural_patterns']:
+                    st.write(f"• {pattern}")
+        
+        with tab4:
+            # Load priority details for this interview
+            with self.db.get_session() as session:
+                interview_obj = session.query(Interview).filter_by(interview_id=interview_id).first()
+                if interview_obj and interview_obj.priorities:
+                    for priority in interview_obj.priorities:
+                        st.markdown(f"**{priority.scope.title()} Priority #{priority.rank}: {priority.theme}**")
+                        
+                        col1, col2 = st.columns([2, 1])
+                        
+                        with col1:
+                            if priority.narrative_elaboration:
+                                st.write(priority.narrative_elaboration)
+                            
+                            if priority.supporting_quotes:
+                                st.markdown("*Supporting quotes:*")
+                                for quote in priority.supporting_quotes:
+                                    st.markdown(f'<div class="quote-card">"{quote}"</div>', 
+                                               unsafe_allow_html=True)
+                        
+                        with col2:
+                            if priority.emotional_intensity:
+                                st.metric("Emotional Intensity", f"{priority.emotional_intensity:.2f}")
+                            if priority.confidence:
+                                st.metric("Confidence", f"{priority.confidence:.2f}")
+                        
+                        st.markdown("---")
 
 
 def main():
-    """Main dashboard application."""
-    # Sidebar navigation
-    st.sidebar.title("Navigation")
+    """Main interactive dashboard application."""
+    # Title and description
+    st.title("🇺🇾 Uruguay Interview Analysis")
+    st.markdown("### 🎯 Interactive Research Platform")
+    st.markdown("*Dynamic exploration of citizen consultation data with advanced filtering and linked visualizations*")
     
-    page = st.sidebar.radio(
-        "Select Dashboard",
-        ["Overview", "Priorities", "Themes", "Interview Detail"]
-    )
+    # Initialize dashboard
+    dashboard = InteractiveDashboard()
     
-    # Load data
     try:
-        df_interviews, df_priorities, df_themes, interviews = load_interview_data()
+        # Load data
+        interview_df, priority_df = dashboard.load_comprehensive_data()
         
-        # Add filters in sidebar
-        st.sidebar.markdown("### Filters")
+        # Create filter sidebar
+        dashboard.create_filter_sidebar(interview_df)
         
-        # Date range filter
-        if not df_interviews.empty and 'date' in df_interviews.columns:
-            date_range = st.sidebar.date_input(
-                "Date Range",
-                value=(df_interviews['date'].min(), df_interviews['date'].max()),
-                key="date_range"
-            )
+        # Apply filters
+        filtered_df = dashboard.apply_filters(interview_df)
         
-        # Department filter
-        if not df_interviews.empty:
-            departments = ["All"] + sorted(df_interviews['department'].unique().tolist())
-            selected_dept = st.sidebar.selectbox("Department", departments)
+        # Main content area
+        col1, col2 = st.columns([3, 1])
+        
+        with col1:
+            # Overview metrics
+            dashboard.show_overview_metrics(interview_df, filtered_df)
             
-            if selected_dept != "All":
-                df_interviews = df_interviews[df_interviews['department'] == selected_dept]
-                df_priorities = df_priorities[df_priorities['department'] == selected_dept]
-                df_themes = df_themes[df_themes['department'] == selected_dept]
-                interviews = [i for i in interviews if i.department == selected_dept]
+        with col2:
+            # Quick stats
+            st.markdown("**Quick Stats:**")
+            st.metric("Unique Departments", filtered_df['department'].nunique())
+            st.metric("Avg Age Range", "50-64" if not filtered_df.empty else "N/A")
         
-        # Display selected page
-        if page == "Overview":
-            show_overview_dashboard(df_interviews, df_priorities, df_themes)
-        elif page == "Priorities":
-            show_priorities_analysis(df_priorities)
-        elif page == "Themes":
-            show_themes_exploration(df_themes)
-        elif page == "Interview Detail":
-            show_interview_detail(interviews)
-            
+        # Main analysis sections
+        st.markdown("---")
+        
+        # Narrative analysis
+        dashboard.create_linked_narrative_analysis(filtered_df)
+        
+        st.markdown("---")
+        
+        # Agency analysis
+        dashboard.create_interactive_agency_analysis(filtered_df)
+        
+        st.markdown("---")
+        
+        # Interview explorer
+        selected_interview = dashboard.create_interview_selector(filtered_df)
+        
+        if selected_interview:
+            st.markdown("---")
+            dashboard.show_detailed_interview_analysis(selected_interview, filtered_df)
+        
+        # Footer
+        st.markdown("---")
+        st.markdown("*🚀 Interactive Dashboard v2.0 - Enhanced Research Platform*")
+        
     except Exception as e:
-        st.error(f"Error loading data: {str(e)}")
-        st.info("Make sure the database is populated with interview data.")
+        st.error(f"Error loading dashboard: {str(e)}")
+        st.exception(e)
 
 
 if __name__ == "__main__":
