+++ src/dashboard/conversation_browser.py
@@ -14,7 +14,7 @@ project_root = Path(__file__).parent.parent.parent
 sys.path.insert(0, str(project_root))
 
 from src.database.connection import get_db
-from src.database.models_enhanced import Interview, Turn, TurnFunctionalAnalysis, TurnContentAnalysis, TurnEmotionalAnalysis
+from src.database.models import Interview, Turn, TurnFunctionalAnalysis, TurnContentAnalysis, TurnEmotionalAnalysis
 
 # Configure page
 st.set_page_config(
